move regex to constant
