+++ app/server/user.server.ts
@@ -1,5 +1,5 @@
 import bcrypt from 'bcryptjs';
-import type { RegisterForm } from '../utils/types.server';
+import type { RegisterForm } from '../shared/utils/types.server';
 import { prisma } from './prisma.server';
 
 export const createUser = async (user: RegisterForm) => {
