+++ app/shared/utils/validators.server.ts
@@ -1,14 +1,16 @@
+import { emailRegex } from '../constant/validator';
+
 export const validateEmail = (email: string): string | undefined => {
   var validRegex =
-    /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/;
+  emailRegex;
   if (!email.length || !validRegex.test(email)) {
     return 'Please enter a valid email address';
   }
 };
 
 export const validatePassword = (password: string): string | undefined => {
-  if (password.length < 5) {
-    return 'Please enter a password that is at least 5 characters long';
+  if (password.length < 8) {
+    return 'Please enter a password that is at least 8 characters long';
   }
 };
